+++ api/src/data/utils.ts
@@ -1,5 +1,4 @@
 import * as AWS from 'aws-sdk';
-import * as csvParser from 'csv-parser';
 import * as fileType from 'file-type';
 import * as admin from 'firebase-admin';
 import * as fs from 'fs';
@@ -7,9 +6,7 @@ import * as Handlebars from 'handlebars';
 import * as nodemailer from 'nodemailer';
 import * as path from 'path';
 import * as puppeteer from 'puppeteer';
-import * as readline from 'readline';
 import * as requestify from 'requestify';
-import { Writable } from 'stream';
 import * as strip from 'strip';
 import * as xlsxPopulate from 'xlsx-populate';
 import {
@@ -130,7 +127,7 @@ export const checkFile = async (file, source?: string) => {
 /**
  * Create AWS instance
  */
-const createAWS = async () => {
+export const createAWS = async () => {
   const AWS_ACCESS_KEY_ID = await getConfig('AWS_ACCESS_KEY_ID');
   const AWS_SECRET_ACCESS_KEY = await getConfig('AWS_SECRET_ACCESS_KEY');
   const AWS_BUCKET = await getConfig('AWS_BUCKET');
@@ -1379,130 +1376,4 @@ export const getErxesSaasDomain = () => {
   return NODE_ENV === 'production'
     ? 'https://erxes.io'
     : 'http://localhost:3500';
-};
-
-const getCsvTotalRowCount = ({
-  filePath,
-  uploadType,
-  s3,
-  params
-}: {
-  filePath?: string;
-  uploadType: string;
-  params?: { Bucket: string; Key: string };
-  s3?: any;
-}): Promise<number> => {
-  return new Promise(resolve => {
-    if (uploadType === 'local' && filePath) {
-      const readSteam = fs.createReadStream(filePath);
-
-      let total = 0;
-
-      const rl = readline.createInterface({
-        input: readSteam,
-        terminal: false
-      });
-
-      rl.on('line', () => total++);
-      rl.on('close', () => resolve(total));
-    } else {
-      s3.selectObjectContent(
-        {
-          ...params,
-          ExpressionType: 'SQL',
-          Expression: 'SELECT COUNT(*) FROM S3Object',
-          InputSerialization: {
-            CSV: {
-              FileHeaderInfo: 'USE',
-              RecordDelimiter: '\n',
-              FieldDelimiter: ','
-            }
-          },
-          OutputSerialization: {
-            CSV: {}
-          }
-        },
-        (_, data) => {
-          // data.Payload is a Readable Stream
-          const eventStream = data.Payload;
-
-          let total = 0;
-
-          // Read events as they are available
-          eventStream.on('data', event => {
-            if (event.Records) {
-              total = event.Records.Payload.toString();
-            }
-          });
-          eventStream.on('end', resolve(total));
-        }
-      );
-    }
-  });
-};
-
-export const importBulkStream = ({
-  fileName,
-  bulkLimit,
-  uploadType,
-  handleBulkOperation
-}: {
-  fileName: string;
-  bulkLimit: number;
-  uploadType: 'S3' | 'local';
-  handleBulkOperation: (rows: any, total: number) => Promise<void>;
-}) => {
-  return new Promise(async (resolve, reject) => {
-    let rows: any = [];
-    let readSteam;
-    let total;
-
-    if (uploadType === 'S3') {
-      const AWS_BUCKET = await getConfig('AWS_BUCKET');
-
-      const s3 = await createAWS();
-
-      const errorCallback = error => {
-        throw new Error(error.code);
-      };
-
-      const params = { Bucket: AWS_BUCKET, Key: fileName };
-
-      total = await getCsvTotalRowCount({ s3, params, uploadType: 's3' });
-
-      readSteam = s3.getObject(params).createReadStream();
-      readSteam.on('error', errorCallback);
-    } else {
-      const filePath: string = `${uploadsFolderPath}/${fileName}`;
-
-      readSteam = fs.createReadStream(filePath);
-      total = await getCsvTotalRowCount({ filePath, uploadType });
-    }
-
-    // exclude column
-    total--;
-
-    const write = (row, _, callback) => {
-      rows.push(row);
-
-      if (rows.length === bulkLimit) {
-        return handleBulkOperation(rows, total).then(() => {
-          rows = [];
-          callback();
-        });
-      }
-
-      return callback();
-    };
-
-    readSteam
-      .pipe(csvParser())
-      .pipe(new Writable({ write, objectMode: true }))
-      .on('finish', () => {
-        handleBulkOperation(rows, total).then(() => {
-          resolve('success');
-        });
-      })
-      .on('error', () => reject('fail'));
-  });
 };
